% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurnThin_function.R
\name{BurnThin}
\alias{BurnThin}
\title{Burn in and thinning of the LERCA fit.}
\usage{
BurnThin(lerca, burn, thin)
}
\arguments{
\item{lerca}{The LERCA fit.}

\item{burn}{The number of initial iterations to discard.}

\item{thin}{The distance of iterations we want to keep.}
}
\description{
Burn in and thinning of the LERCA fit.
}
