% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureCoefs_function.R
\name{ExposureCoefs}
\alias{ExposureCoefs}
\title{Acquire the coefficients as a function of the exposure.}
\usage{
ExposureCoefs(lerca, exp_values)
}
\arguments{
\item{lerca}{The LERCA fit.}

\item{exp_values}{A vector of exposure values.}
}
\description{
LERCA coefficients in each experiment are translated to coefficients as a
function of the exposure.
}
