% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureInclusion_function.R
\name{ExposureInclusion}
\alias{ExposureInclusion}
\title{Acquire the inclusions probabilities as a function of the exposure.}
\usage{
ExposureInclusion(lerca, exp_values)
}
\arguments{
\item{lerca}{The LERCA fit.}

\item{exp_values}{A vector of exposure values where the inclusion
probabilities should be calculated.}
}
\description{
LERCA inclusion indicators in each experiment are translated to inclusion
probabilities as a function of the exposure.
}
