% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetbYvalues_function.R
\name{GetbYvalues}
\alias{GetbYvalues}
\title{Get the bY values that make the ER curve continuous in simulated data when
the X-Y relationship is truly piecewise linear.}
\usage{
GetbYvalues(num_exper, Xrange = NULL, exper_change = NULL, bYX,
  interYexp1 = 0, out_coef, meanC, weights = NULL,
  XY_function = c("linear", "other"))
}
\arguments{
\item{num_exper}{The number of experiments. Should be > 1.}

\item{Xrange}{The range of the exposure range. Specify if exper_change is
NULL.}

\item{exper_change}{The points of experiment change. Should be including the
edges of the exposure range. If left NULL, exper_change will be specified as
equally distanced points on Xrange such that num_exper experiments are
defined.}

\item{bYX}{If XY_function is set to linear, bYX includes the coefficient of
the exposure within each experiment and is of length num_exper. If
XY_function is set to other, bYX is of length 1 and corresponds to the
coefficient in front of the exposure term.}

\item{interYexp1}{The intercept of the outcome model in experiment 1. If not
specified, it defaults to 0.}

\item{out_coef}{A matrix with number of rows equal to the number of
covariates and number of columns equal to the number of experiments
including the outcome model coefficients.}

\item{meanC}{A matrix of equal dimensions to out_coef including the mean of
each covariate in each experiment.}

\item{weights}{The weights to be given to the mean of each experiment in the
calculation of the overall mean. Vector of length equal to the number of
experiments. If left NULL, each experiment is given equal weight.}

\item{XY_function}{String specifying whether the XY relationship is piece-
wise linear (set 'linear'), or a continuous function supplied by the XY_spec
arguement (set 'other').}
}
\description{
Get the bY values that make the ER curve continuous in simulated data when
the X-Y relationship is truly piecewise linear.
}
