% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetER_function.R
\name{GetER}
\alias{GetER}
\title{CER from MCMC of multiple chains.}
\usage{
GetER(dta, cutoffs, coefs, predict_at = NULL, grid_length = 100,
  mean_only = FALSE)
}
\arguments{
\item{dta}{The dataset including columns X (treatment), Y (outcome), and
potential confounders named as C1, C2, ...}

\item{cutoffs}{A 3-dimensional array with dimensions corresponding to chain,
iteration, and number of points in the experiment configuration.}

\item{coefs}{A 4-dimensional array with dimensions corresponding to chain,
iteration, experiment, and variable. Variables are intercept, exposure,
and the potential confounders.}

\item{predict_at}{A vector of the exposure values we wish to predict ER at.
If left NULL, a grid of length grid_length will be created.}

\item{grid_length}{The number of locations we will predict the ER. Defaults
to 100.}

\item{mean_only}{Logical. Set to FALSE if we want the individual ER
predictions. Set to TRUE if we are only interested in the posterior samples
of the mean ER. Defaults to FALSE.}
}
\description{
When multiple chains are ran, we use GetER to get the posterior CER samples.
}
