% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetER_1chain_function.R
\name{GetER_1chain}
\alias{GetER_1chain}
\title{Calculating the exposure response function for one MCMC chain.}
\usage{
GetER_1chain(
  dta,
  cutoffs,
  coefs,
  predict_at = NULL,
  grid_length = 100,
  mean_only = FALSE,
  other_function = NULL
)
}
\arguments{
\item{dta}{The dataset including columns X (treatment), Y (outcome), and potential
confounders named as C1, C2, ...}

\item{cutoffs}{A list of length equal to the number of posterior samples. Each
element of the list is a vector including the current set of cutoff values.}

\item{coefs}{A list of length equal to the number of posterior samples. Each
element of the list is a matrix of the outcome coefficients. The dimension of each
matrix is (number of experiments at current iteration) x (number of potential
confounders). Confounders with inclusion indicators equal to 0, have coefficients
equal to 0. The first two elements of each row of the matrix correspond to intercept
and slope of the exposure.}

\item{predict_at}{A vector of the exposure values we wish to predict ER at. If left
NULL, a grid of length grid_length over the exposure values will be created.}

\item{grid_length}{The number of locations we will predict the ER. Defaults to 100.}

\item{mean_only}{Logical. Set to FALSE if we want the individual ER predictions. Set
to TRUE if we are only interested in the posterior samples of the mean ER. Defaults
to FALSE.}

\item{other_function}{Whether we want to apply a different function to the
predictions. Defaults to NULL. Examples include exp(x), log(x).}
}
\description{
Function that takes the posterior samples of cutoffs, inclusion indicators and
coefficients of the outcome model and returns the posterior mean response for
multiple values of the exposure.
}
