% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JumpOver_function.R
\name{JumpOver}
\alias{JumpOver}
\title{MCMC jump over move}
\usage{
JumpOver(dta, current_cutoffs, current_alphas, current_coefs,
  approx_likelihood = TRUE, cov_cols, omega = 5000, mu_priorY, Sigma_priorY,
  comb_probs = c(0.01, 0.5, 0.99), split_probs = c(0.2, 0.95),
  min_exper_sample = 20, jump_slope_tune = 0.05)
}
\arguments{
\item{dta}{Data frame including the covariates as C1, C2, ..., the exposure
as X and the outcome as Y.}

\item{current_cutoffs}{Numeric of length K. The current values for the
points in the experiment configuraiton.}

\item{current_alphas}{Array of dimensions that correspond to the exposure or
outcome model, the experiment, and potential confounding. Entries are 0/1
corresponding to exclusion/inclusion of the covaraite in the corresponding
model of the experiment.}

\item{current_coefs}{The current coefficients in an array format, with
dimensions corresponding to the exposure/outcome model, the experiments, and
the coefficient (intercept, slope, covariates).}

\item{approx_likelihood}{Logical. If set to true the BIC will be used to
calculate the marginal likelihood. FALSE not supported yet.}

\item{cov_cols}{The indices of the columns including the covariates.}

\item{omega}{The parameter of the BAC prior on inclusion indicators.}

\item{mu_priorY}{Vector of length equal to the number of covariates + 2 with
entries corresponding to the prior mean of the intercept, slope, coefficient
in the outcome model.}

\item{Sigma_priorY}{The normal prior covariance matrix of the parameters in
mu_priorY.}

\item{comb_probs}{When two experiments are combined, comb_probs represents
the probability of alpha = 1 when 0, 1, and 2 corresponding alphas are equal
to 1. Vector of length 3.}

\item{split_probs}{When one experiment is split, split_probs describes the
probability that the alpha of a new experiment is equal to 1, when the alpha
of the current experiment is 0, and when it is 1. Vector of length 2.}

\item{min_exper_sample}{The minimum number of observations within an
experiment. Defaults to 20.}

\item{jump_slope_tune}{The standard deviation of the proposal on the slopes.}
}
\description{
Simultaneous update of the experiment configuration, inclusion indicators,
and slopes that reorder to points in the experiment configuration.
}
