% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JumpOverCoef_function.R
\name{JumpOverCoef}
\alias{JumpOverCoef}
\title{Proposing coefficients for the Jump Pver MCMC move.}
\usage{
JumpOverCoef(current_coefs, prop_cuts, cuts, curr_exper_same, prop_exper_same,
  curr_exper_comb, prop_exper_comb, curr_exper_split, prop_exper_split,
  jump_slope_tune)
}
\arguments{
\item{current_coefs}{The current coefficients of the MCMC. Three dimensional
array with dimensions corresponding to exposure/outcome model, experiment,
and coefficients (intercept, slope, covariates).}

\item{prop_cuts}{Numeric vector with the K locations of the proposed
experiment configuration.}

\item{cuts}{Numeric vector with the K locations of the current experiment
configuration.}

\item{curr_exper_same}{Numeric vector. Indices of experiments in the current
experiment configuration that did not get combined or split. They should be
in the same order as prop_exper_same.}

\item{prop_exper_same}{Numeric vector with the indices of the experiments
in the proposed experiment configuration that did not get combined or split.}

\item{curr_exper_comb}{Numeric vector of length two with entries the indices
of the experiments with respect to the current experiment configuration that
were proposed to be combined.}

\item{prop_exper_comb}{Numeric. Index of the experiment in the proposed
experiment configuration corresponding to the combined experiment.}

\item{curr_exper_split}{Numeric. Index of the experiment in the current
experiment configuration that is proposed to be split.}

\item{prop_exper_split}{Numeric vector of length two. Indices of the
experiments in the proposed experiment configuration that are the result of
a split in an experiment.}

\item{jump_slope_tune}{Numeric. The standard deviation of the proposal on
slopes for the two experiments in the proposed experiment configuration that
are the result of an experiment's split.}
}
\description{
The Jump Over move is accompanied with proposing new slopes that ensure a
continuous ER.
}
