% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JumpWithin_function.R
\name{JumpWithin}
\alias{JumpWithin}
\title{Simultaneous move of experiment, alphas and slopes within experiment.}
\usage{
JumpWithin(
  dta,
  current_cutoffs,
  current_alphas,
  current_coefs,
  cov_cols,
  approx_likelihood = TRUE,
  omega = 5000,
  mu_priorY,
  Sigma_priorY,
  alpha_probs = c(0.01, 0.5, 0.99),
  min_exper_sample = 20
)
}
\arguments{
\item{dta}{Data frame including the covariates as C1, C2, ..., the exposure
as X and the outcome as Y.}

\item{current_cutoffs}{The current values of the experiment configuration.
Vector of length K.}

\item{current_alphas}{The current values of the inclusion indicators. Array
of dimensions 2 (exposure & outcome model), experiments, covariates.}

\item{current_coefs}{The current values of the coefficients. Array of
dimensions 2 (exposure & outcome model), experiments, and coefficients
(intercept, slope, covariates).}

\item{cov_cols}{The indices of the columns including the covariates.}

\item{approx_likelihood}{Logical. If set to true the BIC will be used to
calculate the marginal likelihood. FALSE not supported yet.}

\item{omega}{The omega parameter of the BAC prior.}

\item{mu_priorY}{Vector of length equal to the number of covariates + 2 with
entries corresponding to the prior mean of the intercept, slope, coefficient
in the outcome model.}

\item{Sigma_priorY}{The normal prior covariance matrix of the parameters in
mu_priorY.}

\item{alpha_probs}{The probability that a proposed alpha is equal to 1, when
0, 1, and 2 alphas of the surrounding experiments are equal to 1. Vector of
length 3. Defaults to (0.01, 0.5, 0.99).}

\item{min_exper_sample}{The minimum number of observations within an
experiment. Defaults to 20.}
}
\description{
Proposing a simultaneous move of the experiment configuration, inclusion
indicators and slopes of the exposure in the outcome model while maintaining
the order of the experiment configuration. New values for the slopes are
proposed ensuring continuous ER. We use the likelihood integrating out the
coefficients of the covariates and variance terms.
}
