% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LERCA_function.R
\name{LERCA}
\alias{LERCA}
\title{Local Exposure-Response Confounding Adjustment.}
\usage{
LERCA(dta, chains, Nsims, K, cov_cols, omega = 5000, mu_priorX = NULL,
  Sigma_priorX = NULL, mu_priorY = NULL, Sigma_priorY = NULL,
  alpha_priorX = 0.001, beta_priorX = 0.001, alpha_priorY = 0.001,
  beta_priorY = 0.001, starting_cutoffs = NULL,
  prop_distribution = c("Uniform", "Normal"), normal_percent = 1,
  plot_every = 0, comb_probs = c(0.01, 0.5, 0.99), split_probs = c(0.2,
  0.95), s_upd_probs = c(94/100, 1/100, 5/100), alpha_probs = c(0.01, 0.5,
  0.99), min_exper_sample = 20)
}
\description{
Function that calculates the mean exposure response curve allowing for
differential confounding at different exposure levels.
}
