% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LERCAextend_function.R
\name{LERCAextend}
\alias{LERCAextend}
\title{Extending the LERCA fit.}
\usage{
LERCAextend(dta, extend_Nsims, lerca, plot_every = 0)
}
\arguments{
\item{dta}{The data set including the exposure as X, the outcome as Y, and
all potential confounders as C1, C2, ...}

\item{extend_Nsims}{The number of additional posterior sampels.}

\item{lerca}{The previous LERCA fit.}

\item{plot_every}{At how many iterations we want to plot the posterior
samples of the experiment configuration. Defaults to 0.}
}
\value{
Returns a list with the same elements as LERCA including posterior
samples of the initial and the extended fit.
}
\description{
Acquiring more posterior samples from a LERCA fit.
}
