% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLike_function.R
\name{LogLike}
\alias{LogLike}
\title{Approximate log likelihood of an experiment.}
\usage{
LogLike(D, curr_exper_alphas, curr_coefsY = NULL, X_s_cut, cov_cols)
}
\arguments{
\item{D}{The data set of the current experiment includes covariates,
exposure as 'X' and outcome as 'Y'.}

\item{curr_exper_alphas}{Matrix. Dimensions correspond to exposure/outcome
model and potential confounders. Entries 0/1 represent exlusion/inclusion of
the covariate in each model.}

\item{curr_coefsY}{Numeric of length two. Intercept and slope of the outcome
model. If left NULL, the likelihood is calculated integrating out intercept
and slope along with remaining coefficients.}

\item{X_s_cut}{Numeric. The point in the experiment configuration that
corresponds to the beginning of the current experiment.}

\item{cov_cols}{The indices of the columns in D that correspond to the
potential confounders.}
}
\description{
Calculating the log likelihood based on the BIC approximation.
}
