% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeArrays_function.R
\name{MakeArrays}
\alias{MakeArrays}
\title{Create arrays to save MCMC posterior samples}
\usage{
MakeArrays(
  X = NULL,
  chains,
  Nsims,
  num_exper,
  num_conf,
  omega,
  minX,
  maxX,
  starting_cutoffs,
  starting_alphas,
  starting_coefs,
  starting_vars,
  min_exper_sample = 0
)
}
\arguments{
\item{X}{Numeric vector. Observed exposure values. Can be left NULL if
min_exper_sample is set to 0.}

\item{chains}{The number of separate MCMC chains.}

\item{Nsims}{The number of posterior samples per chain.}

\item{num_exper}{The number of experiments we are allowing.}

\item{num_conf}{The number of potential confounders.}

\item{omega}{The omega of the BAC prior on inclusion indicators.}

\item{minX}{The minimum observed exposure value.}

\item{maxX}{The maximum observed exposure value.}

\item{starting_alphas}{Array with dimensions corresponding to the model
(exposure / outcome), the experiment, and the potential confounders. Entries
0/1 represent exclusion/inclusion of the covariate in the corresponding
model.}

\item{starting_coefs}{Array with the starting values of all coefficients.
Dimensions are: Exposure/Outcome model, chains, experiments, and covariate
(intercept, coefficient of exposure, covariates). The coefficient of
exposure should be NA for the exposure model.}

\item{starting_vars}{Array including the starting values for the residual
variances. Dimensions correspond to: Exposure/Outcome model, chains, and
experiment.}

\item{min_exper_sample}{The minimum number of observations within an
experiment. It will be used to ensure that starting cutoffs are allowed
under the prior specification.}

\item{starting}{cutoffs Matrix with rows corresponding to different chains.
Each row includes K ordered values of MCMC starting cutoffs. If left NULL,
random started values are used.}
}
\description{
Create arrays to save MCMC posterior samples
}
