% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelWeights_function.R
\name{ModelWeights}
\alias{ModelWeights}
\title{Posterior weights of all models in MCMC.}
\usage{
ModelWeights(mcmc, model = c("Outcome", "Exposure"), experiment = 1)
}
\arguments{
\item{mcmc}{The LERCA fit.}

\item{model}{Options are 'Outcome' or 'Exposure'. Defines which model we 
want to look at.}

\item{experiment}{Which experiment to look at. Defaults to 1. Cannot be more
than K + 1. Results will be interpretable for the exposure range in the
intersection of experiments across MCMC iterations.}
}
\description{
Function that calculates and orders the models that were chosen in the MCMC
at least once.
}
