% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLERCA_function.R
\name{PlotLERCA}
\alias{PlotLERCA}
\title{Plotting LERCA results.}
\usage{
PlotLERCA(dta, lerca, ER, variable = NULL, wh_model = NULL, probs,
  inclusion = NULL, coefs = NULL)
}
\arguments{
\item{dta}{The data that we are using with column X for exposure.}

\item{lerca}{The LERCA fit.}

\item{ER}{The estimated ER as acquired using GetER().}

\item{variable}{Which variable's results to be plotted. Intercept is 1,
Exposure is 2, and the covariates are 3 onwards.}

\item{wh_model}{Which model's coefficients and inclusion probabilities to
be plotted. Options are 1, 2 for exposure and outcome model accordingly.}

\item{probs}{The quantiles of the distribution to be plotted as intervals.}

\item{inclusion}{The inclusion probabilities as a function of the exposure
and as acquired using ExposureInclusion().}

\item{coefs}{The coefficients as a function of the exposure as acquired
using ExposureCoefs().}
}
\description{
Estimated ER, cutoffs, coefficients, inclusion probabilities.
}
