% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposeSlopes_function.R
\name{ProposeSlopes}
\alias{ProposeSlopes}
\title{Proposed slope values of Jump Over move.}
\usage{
ProposeSlopes(wh_cut, current_coefs, cuts, sj1star, sj1, sj, sj2)
}
\arguments{
\item{wh_cut}{Integer. Which one of the K points in the experiment
configuration was selected to be moved.}

\item{current_coefs}{The current coefficients in an array format, with
dimensions corresponding to the exposure/outcome model, the experiments, and
the coefficient (intercept, slope, covariates).}

\item{cuts}{The points in the experiment configuration includin the minimum
and maximum values.}

\item{sj1star}{The proposed point in the experiment configuration.}

\item{sj1}{The current point that is proposed to be moved.}

\item{sj}{The point in the current experiment configuration before sj1.}

\item{sj2}{The point in the current experiment configuration after sj1.}
}
\description{
Propose values for the slopes of all experiments in the Jump Over move
ensuring continous ER and unchanged likelihood in experiments that are not
affected.
}
