% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr_function.R
\name{psr}
\alias{psr}
\title{Calculating the PSR of the LERCA fit.}
\usage{
psr(samples)
}
\arguments{
\item{samples}{Matrix. Rows and columns correspond to iterations and chains
accordingly.}
}
\value{
Numeric. Estimate of the potential scale reduction factor. Values
close to 1 indicate MCMC convergence.
}
\description{
Based on the posterior samples of the mean ER estimates, calculate the
potential scale reduction factor to assess MCMC convergence.
}
