% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarToVector_function.R
\name{ScalarToVector}
\alias{ScalarToVector}
\title{Turning variable into vector if scalar.}
\usage{
ScalarToVector(scalar, reps)
}
\arguments{
\item{scalar}{The variable to be turned into a vector if not of length reps.}

\item{reps}{The length that the variable should have.}
}
\value{
A vector of length reps, with scalar repeated until the appropriate
length is achieved.
}
\description{
Turning variable into vector if scalar.
}
