% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDifferentialConfounding_function.R
\name{SimDifferentialConfounding}
\alias{SimDifferentialConfounding}
\title{Simulating data with differential confounding.}
\usage{
SimDifferentialConfounding(N, num_exper, XCcorr, varC, Xrange,
  exper_change = NULL, meanCexp1 = NULL, out_coef, interYexp1 = 0, bYX,
  Ysd = 1, overall_meanC = c("true", "observed"),
  XY_function = c("linear", "other"), XY_spec = NULL)
}
\arguments{
\item{N}{Sample size.}

\item{num_exper}{Number of experiments.}

\item{XCcorr}{A matrix of 2 dimensions. Element ij corresponds to the
correlation of variable Ci with exposure X in experiment j.}

\item{varC}{A vector of length equal to the number of covariates, describing
the variance of each of the C. We assume constant variance of C across
experiments.}

\item{Xrange}{Vector of length 2. These values correspond to the minimum and
the maximum of the uniform distribution from which X is generated.}

\item{meanCexp1}{A vector of length equal to the number of covariates
including the marginal mean of each covariate in experiment 1. The mean of 
the covariates in the rest experiments is specified by the algorithm such
that E(C|X=x) is continuous in x. If equal to NULL, 0 will be specified for
all variables.}

\item{out_coef}{A matrix of two dimensions. Element ij describes the
coefficient of Ci in the outcome model of experiment j.}

\item{interYexp1}{The intercept of the outcome model in experiment 1. The
remaining intercepts are set such that the function E[Y | X = x] is
continuous in x.}

\item{bYX}{If XY_function is set to linear, bYX includes the coefficient of
the exposure within each experiment and is of length num_exper. If
XY_function is set to other, bYX is of length 1 and corresponds to the
coefficient in front of the exposure term.}

\item{Ysd}{A vector including the residual variance of the outcome model in
each experiment. Defaults to 1.}

\item{overall_meanC}{A string equal to 'true' or 'observed'. Defaults to
'true'. This argument controls whether we will use the analytical or the
observed mean of C to ensure continuous ER at the points of the experiment
configuration. If set to 'true', individual simulated data sets do not
necessarily have a continuous ER, but they do in general. If set equal to
the 'observed', each simulated data set has a continuous ER, that might be
slightly different for every dataset. Both choices average to the same true
ER, for a large sample size, or number of replicated data sets.}

\item{XY_function}{String specifying whether the XY relationship is piece-
wise linear (set 'linear'), or a continuous function supplied by the XY_spec
arguement (set 'other').}

\item{XY_spec}{Needs to be specified if XY_function is set to 'other'. It is
the function that specifies the true ER relationship. Defaults to NULL.}

\item{exper_range}{The points of the exposure range Xrange at which
differential confounding might occur.}
}
\description{
Funcition that generates data where different variables are confounders of
the exposure-response relationship based on the exposure range we are
looking at.
}
\examples{
XCcorr <- matrix(c(0.4, 0.2, 0.2, 0), 2, 2)
out_coef <- matrix(c(0.2, 0.3, 0, 0.4), 2, 2)
bYX <- c(0, 0, 1, 2, 3)
sim <- SimDifferentialConfounding(N = 1000, num_exper = 2, XCcorr = XCcorr,
                                  varC = c(1, 1), Xrange = c(0, 10),
                                  exper_change = c(0, 4, 10),
                                  out_coef = out_coef, bYX = bYX)
f <- function(x) {
  return(x ^ 2)
}
sim <- SimDifferentialConfounding(N = 1000, num_exper = 2, XCcorr = XCcorr,
                                  varC = c(1, 1), Xrange = c(0, 10),
                                  exper_change = c(0, 4, 10),
                                  out_coef = out_coef, bYX = bYX,
                                  XY_function = 'other', XY_spec = f)
}
