% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateIntSlope_function.R
\name{UpdateIntSlope}
\alias{UpdateIntSlope}
\title{Intercepts and coefficients of exposure}
\usage{
UpdateIntSlope(
  dta,
  cov_cols,
  current_cutoffs,
  current_coefs,
  current_vars,
  mu_priorY,
  Sigma_priorY
)
}
\arguments{
\item{dta}{A data set including a column of the exposure of interest as X,
the outcome of interest as Y, and all potential confounders as C1, C2, ...}

\item{cov_cols}{The indices of the columns in dta corresponding to the
potential confounders.}

\item{current_cutoffs}{Numeric of length K. The current values for the
points in the experiment configuraiton.}

\item{current_coefs}{The current coefficients of the MCMC. Three dimensional
array with dimensions corresponding to exposure/outcome model, experiment,
and coefficients (intercept, slope, covariates).}

\item{current_vars}{Matrix. Rows correspond to exposure/outcome model, and
columns to experiments. Entries are the current variances.}

\item{mu_priorY}{The mean of the normal prior on the coefficients of the
outcome model. Numeric vector with entries corresponding to intercept, slope
of exposure, and potential covariates.  If left NULL, it is set to 0 for all
parameters.}

\item{Sigma_priorY}{The covariance matrix of the normal prior on the
regression coefficients of the outcome model. If left NULL, it is set to
diagonal with entries 100 ^ 2 for all parameters.}
}
\value{
A matrix with dimensions that correspond to the exposure/outcome
model, and the intercept/slope coefficient.
}
\description{
Updating the outcome model intercepts and coefficients for the exposure.
}
