% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateVariances_function.R
\name{UpdateVariances}
\alias{UpdateVariances}
\title{Exposure and outcome model residual variances}
\usage{
UpdateVariances(dta, current_cutoffs, current_coefs, cov_cols,
  alpha_priorX, beta_priorX, alpha_priorY, beta_priorY)
}
\arguments{
\item{dta}{A data set including a column of the exposure of interest as X,
the outcome of interest as Y, and all potential confounders as C1, C2, ...}

\item{current_cutoffs}{Numeric of length K. The current values for the
points in the experiment configuraiton.}

\item{current_coefs}{The current coefficients of the MCMC. Three dimensional
array with dimensions corresponding to exposure/outcome model, experiment,
and coefficients (intercept, slope, covariates).}

\item{cov_cols}{The indices of the columns in dta corresponding to the
potential confounders.}

\item{alpha_priorX}{The shape parameter of the inverse gamma prior on the
residual variance of the exposure model.}

\item{beta_priorX}{The rate parameter of the inverse gamma prior on the
residual variance of the exposure model.}

\item{alpha_priorY}{The shape parameter of the inverse gamma prior on the
residual variance of the outcome model.}

\item{beta_priorY}{The rate parameter of the inverse gamma prior on the
residual variance of the outcome model.}
}
\value{
Matrix with rows corresponding to the exposure/outcome model and
columns corresponding to the experiment.
}
\description{
Updating the residual variance of the exposure and outcome models within
each experiment.
}
