% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAIC_function.R
\name{WAIC}
\alias{WAIC}
\title{WAIC of LERCA fit using the full conditionals.}
\usage{
WAIC(lerca, dta)
}
\arguments{
\item{lerca}{The LERCA output including cutoffs, coefficients and variances
of the exposure and outcome models.}

\item{dta}{The full data set including the confounders as C1, C2, ..., the
outcome as Y and the exposure as X.}
}
\description{
WAIC of LERCA fit using the full conditionals.
}
