% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCcontinuous_function.R
\name{XCcontinuous}
\alias{XCcontinuous}
\title{Marginal mean of C across experiments to have E(C|X =x) continuous in x.}
\usage{
XCcontinuous(meanCexp1, XCcov, exper_change, meanX, varX)
}
\arguments{
\item{meanCexp1}{A vector of length equal to the number of covariates
including the marginal mean of each covariate in experiment 1.}

\item{XCcov}{A matrix of two dimensions. Element ij describes the covariance
of variable Ci with X in experiment j.}

\item{exper_change}{A vector including the points sbar that specify the
experiments. exper_change includes the minimum and maximum of the exposure
range.}

\item{meanX}{A vector including the mean of X in each experiment.}

\item{varX}{A vector including the variance of X in each experiment.}
}
\description{
Marginal mean of C across experiments to have E(C|X =x) continuous in x.
}
